+++ src/vue/modules/movements-top-bar/index.vue
@@ -94,7 +94,7 @@ import TopBar from '@/vue/common/TopBar'
 import Drawer from '@/vue/common/Drawer'
 import SelectField from '@/vue/fields/SelectField'
 import TransferForm from '@/vue/forms/TransferForm'
-
+import { ErrorHandler } from '@/js/helpers/error-handler'
 import { mapActions, mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex'
 import { ALL_VALUE } from '@/js/const/select-field-default-values.const'
@@ -103,6 +103,7 @@ const EVENTS = {
   assetCodeUpdated: 'asset-code-updated',
   movementsUpdateRequired: 'movements-update-required',
   showNoDataMessage: 'show-no-data-message',
+  showLoadingErrorMessage: 'show-loading-error-message',
 }
 
 const ASSET_POLICIES_STR = {
@@ -215,6 +216,14 @@ export default {
     getBalance () {
       const balance = +this.accountBalanceByCode(this.assetCode).balance
       this.isHaveBalance = balance > 0
+    },
+    async loadAccountBalances () {
+      try {
+        await this.loadAccountBalancesDetails()
+      } catch (error) {
+        this.$emit(EVENTS.showLoadingErrorMessage)
+        ErrorHandler.processWithoutFeedback(error)
+      }
     },
   },
 }