+++ src/vue/pages/atomic-swaps/AtomicSwapsExplore.vue
@@ -28,6 +28,22 @@
       </div>
     </template>
 
+    <template v-else-if="isLoadingFailed">
+      <p>{{ 'atomic-swaps-explore.loading-error-msg' | globalize }}</p>
+    </template>
+
+    <template v-else-if="!list.length && isLoading">
+      <div class="atomic-swaps-explore__list">
+        <div
+          class="atomic-swaps-explore__list-item-wrp"
+          v-for="item in 5"
+          :key="item"
+        >
+          <atomic-swap-card-skeleton />
+        </div>
+      </div>
+    </template>
+
     <template v-else-if="!list.length && !isLoading">
       <no-data-message
         class="atomic-swaps-explore__no-data-message"
@@ -115,6 +131,7 @@ export default {
   data () {
     return {
       isLoading: false,
+      isLoadingFailed: false,
       list: [],
       isBuyFormDrawerShown: false,
       isAtomicSwapDetailsDrawerShown: false,
@@ -167,6 +184,7 @@ export default {
           filter: filter,
         })
       } catch (error) {
+        this.isLoadingFailed = true
         ErrorHandler.processWithoutFeedback(error)
       }